Fix renaming mistake
